Safari-611.1.10/LayoutTests/fast/css/deferred-parsing/nth-of-type-expected.txt

This tests deferred parsing of CSS :nth-of-type style rules where some are invalid.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



PASS document.styleSheets[0].cssRules.length is 4
PASS internals.deferredStyleRulesCount(document.styleSheets[0]) is 4
PASS internals.deferredStyleRulesCount(document.styleSheets[0]) is 2
PASS successfullyParsed is true

TEST COMPLETE
